# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/cooks": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`).",
        "operationId": "CooksController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "schema": {
              "example": 3.8833,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "example": -77.0319,
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "Search radius in meters",
            "schema": {
              "default": 10000,
              "example": 10000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active cooks (optional proximity: lat, lng, radius meters)",
        "tags": [
          "cooks"
        ]
      },
      "post": {
        "operationId": "CooksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create cook profile for an existing user (admin only)",
        "tags": [
          "cooks"
        ]
      }
    },
...
    "/api/cooks/{id}": {
      "get": {
        "operationId": "CooksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get cook profile (exact coordinates only for owner/admin; public sees city label)",
        "tags": [
          "cooks"
        ]
      },
      "patch": {
        "operationId": "CooksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update cook profile (owner or admin; isActive is admin-only)",
        "tags": [
          "cooks"
        ]
      }
    },
    "/api/products": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`). Query `tags=seafood,shrimp` uses AND semantics. Pass `lat`+`lng` (+ optional `radius`) to filter by cook proximity. Product `images` are `{ id, url }` objects.",
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive name search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cookId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isAvailable",
            "required": false,
            "in": "query",
            "description": "Defaults to true (only available products in catalog)",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "availability",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "available",
                "made_to_order"
              ]
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Comma-separated tags; AND semantics (product must include all). Normalized lowercase.",
            "schema": {
              "example": "seafood,shrimp",
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "description": "With lng+radius: proximity filter on GET /products",
            "schema": {
              "example": 3.8833,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "example": -77.0319,
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "Radius in meters",
            "schema": {
              "example": 10000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List products (search, category, cook, price, availability, tags AND, optional lat/lng/radius)",
        "tags": [
          "products"
        ]
      },
      "post": {
        "description": "Supports `preparationTimeHours` and optional `tags` (must exist in GET /tags catalog; create tags via POST /tags as admin). Upload images via POST /products/:id/images after create (max 5).",
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create product (cook: own catalog; admin: must pass cookId)",
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/products/{id}": {
      "get": {
        "description": "Images are `{ id, url }` (upload via POST /products/:id/images).",
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get product by id",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product (owner cook or admin; isActive admin-only)",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete product (owner cook or admin)",
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/categories": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`). Optional `search` matches name or description.",
        "operationId": "CategoriesController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial text match (fields depend on the resource).",
            "schema": {
              "example": "maría",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List active categories",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create category (admin)",
        "tags": [
          "categories"
        ]
      }
    },
...
    "/api/orders": {
      "post": {
        "description": "Single-cook order. Stock is reserved for `available` products.\n\n**Delivery (required)** — copied onto the order at create time (profile changes later do not affect this order):\n\n- `source: CUSTOMER_PROFILE` — uses `PATCH /users/me/delivery-information`\n- `source: CUSTOM` — requires `location` (GeoJSON Point `[lng, lat]`) and `address`; `additionalInformation` optional",
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              },
              "examples": {
                "customerProfile": {
                  "summary": "Use saved customer delivery",
                  "value": {
                    "cookId": "507f1f77bcf86cd799439011",
                    "items": [
                      {
                        "productId": "507f1f77bcf86cd799439012",
                        "quantity": 2
                      }
                    ],
                    "delivery": {
                      "source": "CUSTOMER_PROFILE"
                    },
                    "paymentMethod": "nequi"
                  }
                },
                "customDelivery": {
                  "summary": "Custom address for this order",
                  "value": {
                    "cookId": "507f1f77bcf86cd799439011",
                    "items": [
                      {
                        "productId": "507f1f77bcf86cd799439012",
                        "quantity": 2
                      }
                    ],
                    "delivery": {
                      "source": "CUSTOM",
                      "location": {
                        "type": "Point",
                        "coordinates": [
                          -77.0319,
                          3.8833
                        ]
                      },
                      "address": "Calle 5 #10-20",
                      "additionalInformation": "Casa azul, next to the bakery"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created order including delivery snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create order with delivery snapshot",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "description": "**Customers (spec 003)** — order history scoped to the JWT user id (never pass `customerId`).\nSupports cursor pagination: `?limit=20&cursor=...` (default 20, max 100).\n\nResponse for customers:\n`{ data: [{ id, status, paymentStatus, total, createdAt }], pagination: { nextCursor, hasMore } }`\n\n**Cooks / admins** — same pagination shape; cooks only see orders for their cook profile (full order objects in `data`).",
        "operationId": "OrdersController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated order list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOrdersListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List orders for the authenticated user",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/admin/stats": {
      "get": {
        "operationId": "AdminController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Basic dashboard stats (cooks, customers, products, orders, sales)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/customers": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`). Optional `search` matches name or email.",
        "operationId": "AdminController_listCustomers",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial text match (fields depend on the resource).",
            "schema": {
              "example": "maría",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customers",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/customers/{id}": {
      "patch": {
        "operationId": "AdminController_setCustomerActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate / deactivate a customer",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/cooks": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`). Optional `search` matches display name, bio, location, specialties, WhatsApp, or linked user name/email.",
        "operationId": "AdminController_listCooks",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial text match (fields depend on the resource).",
            "schema": {
              "example": "maría",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all cooks (including inactive)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/cooks/{id}": {
      "patch": {
        "operationId": "AdminController_setCookActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate / deactivate a cook",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/products": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`).",
        "operationId": "AdminController_listProducts",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Review all products (including inactive)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/products/{id}": {
      "patch": {
        "operationId": "AdminController_setProductActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate / deactivate a product",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/orders": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`). Optional `search` matches order number, status, payment method, or customer name/email.",
        "operationId": "AdminController_listOrders",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial text match (fields depend on the resource).",
            "schema": {
              "example": "maría",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Review all orders",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/categories": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`). Optional `search` matches name or description.",
        "operationId": "AdminController_listCategories",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial text match (fields depend on the resource).",
            "schema": {
              "example": "maría",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all categories (including inactive)",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Market Petronio API",
    "description": "Marketplace API for cooks and customers (Pacific region).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCookDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Existing user id that will own this cook profile",
            "example": "507f1f77bcf86cd799439011"
          },
          "displayName": {
            "type": "string",
            "example": "María Rodríguez"
          },
          "bio": {
            "type": "string",
            "example": "Traditional cook from the Pacific region."
          },
          "specialties": {
            "example": [
              "Seafood",
              "Coconut candies",
              "Tamales"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publicLocation": {
            "type": "string",
            "example": "Buenaventura, Valle del Cauca"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude (GeoJSON order)",
            "example": -77.0319
          },
          "latitude": {
            "type": "number",
            "description": "Latitude (GeoJSON order)",
            "example": 3.8833
          },
          "paymentMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodDto"
            }
          },
          "contactWhatsApp": {
            "type": "string",
            "example": "3001234567"
          }
        },
        "required": [
          "userId",
          "displayName",
          "publicLocation",
          "longitude",
          "latitude"
        ]
      },
      "UpdateCookDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publicLocation": {
            "type": "string"
          },
          "longitude": {
            "type": "number"
          },
          "latitude": {
            "type": "number"
          },
          "paymentMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodDto"
            }
          },
          "contactWhatsApp": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "description": "Admin only"
          }
        }
      },
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          },
          "categoryId": {
            "type": "string"
          },
          "availability": {
            "type": "string",
            "enum": [
              "available",
              "made_to_order"
            ]
          },
          "preparationTimeHours": {
            "type": "number",
            "example": 24,
            "description": "Preparation time in hours"
          },
          "minimumOrderQuantity": {
            "type": "number"
          },
          "tags": {
            "example": [
              "seafood",
              "shrimp"
            ],
            "description": "Must already exist in the global tag catalog (POST /tags). Normalized to lowercase; max 10; unique",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isAvailable": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean",
            "description": "Admin soft-deactivate"
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Prepared meals"
          },
          "description": {
            "type": "string",
            "example": "Ready-to-eat traditional dishes"
          },
          "image": {
            "type": "string",
            "example": "https://res.cloudinary.com/demo/image/upload/category.jpg"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}