# src/admin/admin.controller.ts
import { Body, Controller, Get, Param, Patch, Query } from '@nestjs/common';
import { ApiBearerAuth, ApiOperation, ApiTags } from '@nestjs/swagger';
import { Roles } from '../common/decorators/roles.decorator';
import { SearchQueryDto } from '../common/dto/search-query.dto';
import { CursorPaginationQueryDto } from '../common/pagination/cursor-pagination.dto';
import { Role } from '../common/enums/role.enum';
import { AdminService } from './admin.service';
import { SetActiveDto } from './dto/set-active.dto';

@ApiTags('admin')
@ApiBearerAuth()
@Roles(Role.Admin)
@Controller('admin')
export class AdminController {
  constructor(private readonly adminService: AdminService) {}

  @Get('stats')
  @ApiOperation({
    summary:
      'Basic dashboard stats (cooks, customers, products, orders, sales)',
  })
  stats() {
    return this.adminService.getStats();
  }

  @Get('customers')
  @ApiOperation({
    summary: 'List customers',
    description:
      'Cursor pagination (`limit`, `cursor`). Optional `search` matches name or email.',
  })
  listCustomers(@Query() query: SearchQueryDto) {
    return this.adminService.listCustomers(query);
  }

  @Patch('customers/:id')
  @ApiOperation({ summary: 'Activate / deactivate a customer' })
  setCustomerActive(@Param('id') id: string, @Body() dto: SetActiveDto) {
    return this.adminService.setCustomerActive(id, dto.isActive);
  }

  @Get('cooks')
  @ApiOperation({
    summary: 'List all cooks (including inactive)',
    description:
      'Cursor pagination (`limit`, `cursor`). Optional `search` matches display name, bio, location, specialties, WhatsApp, or linked user name/email.',
  })
  listCooks(@Query() query: SearchQueryDto) {
    return this.adminService.listCooks(query);
  }

  @Patch('cooks/:id')
  @ApiOperation({ summary: 'Activate / deactivate a cook' })
  setCookActive(@Param('id') id: string, @Body() dto: SetActiveDto) {
    return this.adminService.setCookActive(id, dto.isActive);
  }

  @Get('products')
  @ApiOperation({
    summary: 'Review all products (including inactive)',
    description: 'Cursor pagination (`limit`, `cursor`).',
  })
  listProducts(@Query() query: CursorPaginationQueryDto) {
    return this.adminService.listProducts(query);
  }

  @Patch('products/:id')
  @ApiOperation({ summary: 'Activate / deactivate a product' })
  setProductActive(@Param('id') id: string, @Body() dto: SetActiveDto) {
    return this.adminService.setProductActive(id, dto.isActive);
  }

  @Get('orders')
  @ApiOperation({
    summary: 'Review all orders',
    description:
      'Cursor pagination (`limit`, `cursor`). Optional `search` matches order number, status, payment method, or customer name/email.',
  })
  listOrders(@Query() query: SearchQueryDto) {
    return this.adminService.listOrders(query);
  }

  @Get('categories')
  @ApiOperation({
    summary: 'List all categories (including inactive)',
    description:
      'Cursor pagination (`limit`, `cursor`). Optional `search` matches name or description.',
  })
  listCategories(@Query() query: SearchQueryDto) {
    return this.adminService.listCategories(query);
  }
}